requirements.txt
sv-ttk
pymupdf
pillow
pandas
openpyxl
google-genai
python-dotenv
